# docs/foundry-openapi-connector.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HomeCare Uhie Foundry Tools",
    "version": "3.0.0",
    "description": "Role-isolated Foundry tool API. Pass actingRole and actingEmployeeCode from signed-in chat context on every call. Caregiver and admin tools are on separate paths. All responses use HTTP 200 — check success and message in the body."
  },
  "servers": [
    {
      "url": "https://backend-home-care-scheduler-api.vercel.app/api/uhie/tools",
      "description": "Production Vercel Foundry tools"
    }
  ],
  "security": [{ "apiKeyHeader": [] }],
...
  "paths": {
...
    "/admin/schedules/validate": {
      "post": {
        "tags": ["admin-schedules"],
        "operationId": "validateScheduleAssignment",
        "summary": "Validate a caregiver can be assigned to a client shift",
        "description": "Optional step before assignSchedule.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScheduleAssignmentBody" }
            }
          }
        },
        "responses": { "200": { "description": "Always HTTP 200" } }
      }
    },
...
    "/admin/schedules/{scheduleId}/reassign": {
      "put": {
        "tags": ["admin-schedules"],
        "operationId": "reassignSchedule",
        "summary": "Reassign an existing shift to another caregiver",
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["employeeCode", "actingRole", "actingEmployeeCode"],
                "properties": {
                  "employeeCode": {
                    "type": "string",
                    "description": "New caregiver employeeCode"
                  },
                  "actingRole": { "type": "string", "enum": ["admin"] },
                  "actingEmployeeCode": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Always HTTP 200" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
    "schemas": {
...
      "ScheduleAssignmentBody": {
        "type": "object",
        "required": [
          "actingRole",
          "actingEmployeeCode",
          "clientCode",
          "employeeCode",
          "date",
          "startTime",
          "endTime"
        ],
        "properties": {
          "actingRole": { "type": "string", "enum": ["admin"] },
          "actingEmployeeCode": { "type": "string" },
          "clientCode": { "type": "string" },
          "employeeCode": { "type": "string", "description": "Caregiver to assign" },
          "date": { "type": "string", "format": "date" },
          "startTime": { "type": "string" },
          "endTime": { "type": "string" },
          "notes": { "type": "string" }
        }
      }
    }
  }
}